rename select to select-solution
