+++ app/pages/sections/catalogue-solutions/select-solution/contextCreator.js
