+++ app/pages/sections/catalogue-solutions/select-solution/controller.js
